improve re-connect loop
