+++ session/sessionstate.go
@@ -5,7 +5,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"github.com/qlik-oss/gopherciser/globals/constant"
 	"net/http"
 	"sync"
 	"time"
@@ -665,51 +664,41 @@ func (state *State) Reconnect() error {
 		})
 	}
 
+	// todo make reconnect scheme configurable
 	if state == nil || state.reconnect.reconnectFunc == nil {
 		state.reconnect.err = nil
-		return nil
+		return nil // we should not try to reconnect
 	}
 
-	var err error // todo make reconnect scheme configurable
+reconnectLoop:
 	for _, waitTime := range []int{0, 1, 10, 30, 60} {
 		<-time.After(time.Duration(waitTime) * time.Second)
 
 		reConnectActionState := &action.State{}
 
 		if _, err := state.reconnect.reconnectFunc(); err != nil {
-			return errors.Wrap(err, "Failed connecting to sense server")
+			state.reconnect.err = errors.Wrap(err, "Failed connecting to sense server")
+			continue reconnectLoop
 		}
 
 		if err := state.SetupChangeChan(); err != nil {
-			return errors.Wrap(err, "failed to setup change channel")
+			state.reconnect.err = errors.Wrap(err, "failed to setup change channel")
+			continue reconnectLoop
 		}
 
 		upLink := state.Connection.Sense()
 
-		connectMsgChan := upLink.Global.SessionMessageChannel(constant.EventTopicOnConnected)
-		defer upLink.Global.CloseSessionMessageChannel(connectMsgChan)
-
-		if err := state.SendRequest(reConnectActionState, func(ctx context.Context) error {
-			activeDoc, err := upLink.Global.GetActiveDoc(ctx)
-			if err != nil {
-				return errors.WithStack(NoActiveDocError{Err: err})
-			} else if activeDoc == nil {
-				return errors.WithStack(NoActiveDocError{Msg: "No Active doc found on reconnect."})
-			}
-
-			return upLink.SetCurrentApp(activeDoc.GenericId, activeDoc)
-		}); err != nil {
-			return err
+		doc, err := state.GetActiveDoc(reConnectActionState, upLink)
+		if err != nil {
+			state.reconnect.err = errors.WithStack(err)
+			break reconnectLoop // no active doc, don't try re connecting again
 		}
 
-		select {
-		case <-connectMsgChan:
-		case <-state.BaseContext().Done():
-			return reConnectActionState.Errors()
+		// set active doc as current app
+		if err := upLink.SetCurrentApp(doc.GenericId, doc); err != nil {
+			state.reconnect.err = errors.WithStack(err)
+			break reconnectLoop
 		}
-
-		// todo
-		//openApp.getSheetList(state, reConnectActionState, upLink)
 
 		var wg sync.WaitGroup
 		// re-subscribe to objects
@@ -725,21 +714,18 @@ func (state *State) Reconnect() error {
 		wg.Wait()
 
 		if reConnectActionState.Failed {
-			return reConnectActionState.Errors()
+			state.reconnect.err = reConnectActionState.Errors()
 		}
 
-		//err = state.reconnect.reconnectFunc(subscribedObjects)
-
-		state.reconnect.err = err
-		switch errors.Cause(err).(type) {
+		switch errors.Cause(state.reconnect.err).(type) {
 		case nil:
-			return nil
+			return nil // sucessful re-connect
 		case NoActiveDocError:
-			return errors.Wrap(err, "Reconnect error") // connected but session is gone, don't retry
+			break reconnectLoop // invalid doc, don't try more re-connects
 		}
 	}
 
-	return errors.Wrap(err, "Reconnect error")
+	return errors.Wrap(state.reconnect.err, "Reconnect error")
 }
 
 // SetReconnectFunc sets current app re-connect function
@@ -753,73 +739,4 @@ func (state *State) SetReconnectFunc(f func() (string, error)) {
 // GetReconnectError from latest finished reconnect attempt
 func (state *State) GetReconnectError() error {
 	return state.reconnect.err
-}
-
-func (state *State) SetupChangeChan() error {
-	if state == nil {
-		return errors.New("use of nil state")
-	}
-
-	if state.Connection == nil {
-		return errors.New("connection is nil")
-	}
-
-	if state.Connection.Sense() == nil {
-		return errors.New("sense uplink is nil")
-	}
-
-	if state.Connection.Sense().Global == nil {
-		return errors.New("uplink global is nil")
-	}
-
-	changeChan := state.Connection.Sense().Global.ChangeListsChannel(true)
-	go func() {
-		for {
-			select {
-			case cl, ok := <-changeChan:
-				if !ok {
-					return
-				}
-
-				if len(cl.Changed) > 0 {
-					state.LogEntry.LogInfo("Pushed ChangedList", fmt.Sprintf("%v", cl.Changed))
-				}
-
-				if len(cl.Closed) > 0 {
-					state.LogEntry.LogInfo("Pushed ClosedList", fmt.Sprintf("%v", cl.Closed))
-				}
-
-				state.TriggerEvents(state.CurrentActionState, cl.Changed, cl.Closed)
-			case <-state.BaseContext().Done():
-				return
-			}
-
-		}
-	}()
-
-	return nil
-}
-
-// GetSheetList create and update sheetlist session object if not existing
-func (state *State) GetSheetList(actionState *action.State, uplink *enigmahandlers.SenseUplink) {
-	sheetList, err := uplink.CurrentApp.GetSheetList(state, actionState)
-	if err != nil {
-		actionState.AddErrors(err)
-		return
-	}
-
-	slLayout := sheetList.Layout()
-	if slLayout == nil {
-		actionState.AddErrors(errors.New("sheetlist layout is nil"))
-		return
-	}
-
-	if state.LogEntry.ShouldLogDebug() &&
-		slLayout.AppObjectList != nil &&
-		slLayout.AppObjectList.Items != nil {
-
-		for _, v := range slLayout.AppObjectList.Items {
-			state.LogEntry.LogDebugf("Sheet<%s> found", v.Info.Id)
-		}
-	}
 }
